// UI-Enhanced/src/types/api.ts
export interface LoginRequest {
  email: string;
  password: string;
}